Split search form out of widget component so it can be reused in the main search area
